+++ src/App.tsx
@@ -1,13 +1,10 @@
 import { ChakraProvider } from "@chakra-ui/react";
-import { ScheduleProvider } from "./ScheduleContext.tsx";
-import { ScheduleTables } from "./ScheduleTables.tsx";
+import { ScheduleTables } from "./ScheduleTables/ScheduleTables.tsx";
 
 function App() {
 	return (
 		<ChakraProvider>
-			<ScheduleProvider>
-				<ScheduleTables />
-			</ScheduleProvider>
+			<ScheduleTables />
 		</ChakraProvider>
 	);
 }
