--- src/ScheduleContext.tsx
@@ -1,39 +1,0 @@
-import type React from "react";
-import {
-	createContext,
-	type PropsWithChildren,
-	useContext,
-	useState,
-} from "react";
-import dummyScheduleMap from "./dummyScheduleMap.ts";
-import type { Schedule } from "./types.ts";
-
-interface ScheduleContextType {
-	schedulesMap: Record<string, Schedule[]>;
-	setSchedulesMap: React.Dispatch<
-		React.SetStateAction<Record<string, Schedule[]>>
-	>;
-}
-
-const ScheduleContext = createContext<ScheduleContextType | undefined>(
-	undefined,
-);
-
-export const useScheduleContext = () => {
-	const context = useContext(ScheduleContext);
-	if (context === undefined) {
-		throw new Error("useSchedule must be used within a ScheduleProvider");
-	}
-	return context;
-};
-
-export const ScheduleProvider = ({ children }: PropsWithChildren) => {
-	const [schedulesMap, setSchedulesMap] =
-		useState<Record<string, Schedule[]>>(dummyScheduleMap);
-
-	return (
-		<ScheduleContext.Provider value={{ schedulesMap, setSchedulesMap }}>
-			{children}
-		</ScheduleContext.Provider>
-	);
-};
